add LRU limitation to liveness cacehing
